+++ src/auth_code.rs
@@ -107,7 +107,7 @@ impl OAuthClient for AuthCodeSpotify {
         let mut data = Form::new();
         let oauth = self.get_oauth();
         let scopes = oauth
-            .scope
+            .scopes
             .clone()
             .into_iter()
             .collect::<Vec<_>>()
@@ -177,8 +177,8 @@ impl AuthCodeSpotify {
     pub fn get_authorize_url(&self, show_dialog: bool) -> ClientResult<String> {
         let mut payload: HashMap<&str, &str> = HashMap::new();
         let oauth = self.get_oauth();
-        let scope = oauth
-            .scope
+        let scopes = oauth
+            .scopes
             .clone()
             .into_iter()
             .collect::<Vec<_>>()
@@ -186,7 +186,7 @@ impl AuthCodeSpotify {
         payload.insert(headers::CLIENT_ID, &self.get_creds().id);
         payload.insert(headers::RESPONSE_TYPE, headers::RESPONSE_CODE);
         payload.insert(headers::REDIRECT_URI, &oauth.redirect_uri);
-        payload.insert(headers::SCOPE, &scope);
+        payload.insert(headers::SCOPE, &scopes);
         payload.insert(headers::STATE, &oauth.state);
 
         if show_dialog {